изменил отступы для вертикального скрола браузера, убрал логический false для БЭМ валидатора
