+++ src/components/App/App.css
@@ -15,7 +15,6 @@
 
 .page {
   margin: auto;
-  min-width: 1140px;
   min-height: 100vh;
   display: flex;
   flex-direction: column;
@@ -50,7 +49,7 @@
   letter-spacing: -.04em;
 }
 
-@media (max-width: 1139px) {
+@media (max-width: 1200px) {
   .page {
     min-width: 320px;
   }
@@ -66,7 +65,7 @@
   }
 }
 
-@media (max-width: 767px) {
+@media (max-width: 430px) {
   .title {
     font-size: 30px;
     line-height: 1.26;
