+++ src/components/Form/Form.css
@@ -87,7 +87,7 @@
   margin: 158.59px 0 0;
 }
 
-.form__link {
+.link {
   text-decoration: none;
   color: #A0A0A0;
   text-align: center;
@@ -96,11 +96,11 @@
   line-height: 1.21;
 }
 
-.form__link:hover {
+.link:hover {
   opacity: .7;
 }
 
-.form__text {
+.link__text {
   margin: 0 0 0 6px;
   color: #FF6838;
 }
@@ -123,7 +123,7 @@
     width: 230px;
   }
 
-  .form__link,
+  .link,
   .form__button {
     font-size: 12px;
   }
