+++ src/components/Header/Header.css
@@ -7,7 +7,7 @@
   background-color: #465DFF;
 }
 
-@media (max-width: 1139px) {
+@media (max-width: 1200px) {
   .header {
     padding: 0 30px;
   }
