+++ src/components/Login/Login.css
@@ -6,7 +6,7 @@
   padding: 70px 0;
 }
 
-@media (max-width: 1139px) {
+@media (max-width: 1200px) {
   .login {
     padding: 232px 0 0;
   }
