+++ src/components/Movies/Movies.css
@@ -2,7 +2,7 @@
   padding: 0 calc((100% - 1280px + 140px)/2);
 }
 
-@media (max-width: 1139px) {
+@media (max-width: 1200px) {
   .movies {
     padding: 0 30px;
   }
