+++ src/components/Popup/Popup.jsx
@@ -7,9 +7,9 @@ const modalRoot = document.querySelector('#modals');
 
 export default function Popup({ isOpen, handleTogglePopup }) {
   usePopupClose(isOpen, handleTogglePopup);
-  
+
   return createPortal(
-    <div className={`popup ${isOpen && 'popup_opened'}`}>
+    <div className={`popup ${isOpen ? 'popup_opened' : ''}`}>
       <SideBar isOpen={isOpen} handleTogglePopup={handleTogglePopup} />
     </div>,
     modalRoot,
