+++ src/components/Register/Register.css
@@ -6,7 +6,7 @@
   padding: 70px 0;
 }
 
-@media (max-width: 1139px) {
+@media (max-width: 1200px) {
   .register {
     padding: 232px 0 0;
   }
